+++ pynzb/__init__.py
@@ -1,6 +1,5 @@
 # flake8: noqa: F401
 from pynzb.etree_nzb import ETreeNZBParser
-from pynzb.expat_nzb import ExpatNZBParser
 
 # Try to import lxml parser, but don't fail if it's not available
 try: